% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rmst-datasets.R
\docType{data}
\name{aft_weibull_L24_n200}
\alias{aft_weibull_L24_n200}
\title{Simulated dataset: AFT Weibull, L = 24, n = 200}
\format{
A data frame with 200 rows and 5 variables.
}
\description{
Columns: time, status, arm, age, gender.
}
\examples{
\dontrun{
data(aft_weibull_L24_n200)
mean(aft_weibull_L24_n200$status == 0)
}
}
\keyword{datasets}
