% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen-interface.R
\name{covar_binary}
\alias{covar_binary}
\title{Binary covariate definition}
\usage{
covar_binary(name, p = 0.5)
}
\arguments{
\item{name}{Column name in the generated data.}

\item{p}{Probability of value 1. Default \code{0.5}.}
}
\value{
A named list suitable as one element of the \code{covariates} argument.
}
\description{
Builds a binary (Bernoulli) covariate definition list.
}
\examples{
covar_binary("female", p = 0.52)

}
\seealso{
\code{\link{covar_continuous}}, \code{\link{covar_categorical}},
\code{\link{rmst.sim}}
}
