% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent_censoring_analytical_app.R
\name{DC.power.analytical.app}
\alias{DC.power.analytical.app}
\title{Analyze Power for RMST Model with Dependent Censoring (Analytic)}
\usage{
DC.power.analytical.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  dep_cens_status_var,
  sample_sizes,
  linear_terms = NULL,
  L,
  alpha = 0.05
)
}
\arguments{
\item{sample_sizes}{A numeric vector of sample sizes \emph{per arm} to calculate power for.}

\item{alpha}{The significance level (Type I error rate).}
}
\value{
A list containing results.
}
\description{
Performs power analysis for an RMST model with multiple censoring causes.
}
\keyword{internal}
