% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent_censoring_analytical_app.R
\name{DC.ss.analytical.app}
\alias{DC.ss.analytical.app}
\title{Find Sample Size for RMST Model with Dependent Censoring (Analytic)}
\usage{
DC.ss.analytical.app(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  dep_cens_status_var,
  target_power,
  linear_terms = NULL,
  L,
  alpha = 0.05,
  n_start = 50,
  n_step = 25,
  max_n_per_arm = 2000
)
}
\arguments{
\item{target_power}{A single numeric value for the desired power.}

\item{alpha}{The significance level (Type I error rate).}

\item{n_start}{The starting sample size \emph{per arm} for the search.}

\item{n_step}{The increment in sample size at each step of the search.}

\item{max_n_per_arm}{The maximum sample size \emph{per arm} to search up to.}
}
\value{
A list containing results.
}
\description{
Calculates the required sample size for a target power.
}
\keyword{internal}
