% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_gam.R
\name{design_rmst_gam_ss}
\alias{design_rmst_gam_ss}
\title{Find Sample Size for a Multiplicative Stratified RMST Model (Analytic)}
\usage{
design_rmst_gam_ss(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  strata_var,
  target_power,
  linear_terms = NULL,
  tau,
  alpha = 0.05,
  n_start = 50,
  n_step = 25,
  max_n_per_arm = 2000
)
}
\description{
Calculates the required sample size for a target power using an
analytic formula based on the methods of Wang et al. (2019).
}
\note{
This function is analytic and uses a log-linear approximation.
}
