% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ipcw.R
\name{design_rmst_ipcw_ss}
\alias{design_rmst_ipcw_ss}
\title{Find Sample Size for a Linear RMST Model (Analytic)}
\usage{
design_rmst_ipcw_ss(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  target_power,
  linear_terms = NULL,
  tau,
  alpha = 0.05,
  n_start = 50,
  n_step = 25,
  max_n_per_arm = 2000
)
}
\arguments{
\item{pilot_data}{A data.frame with pilot study data.}

\item{time_var, status_var, arm_var}{Strings for column names.}

\item{target_power}{A single numeric value for the target power.}

\item{linear_terms}{Optional vector of covariates for the linear model.}

\item{tau}{The truncation time for RMST.}

\item{alpha}{The significance level. Default is 0.05.}

\item{n_start}{Starting sample size per arm for the search.}

\item{n_step}{Increment for the sample size search.}

\item{max_n_per_arm}{Maximum sample size per arm to try.}
}
\value{
A list containing \code{results_data}, \code{results_plot}, and \code{results_summary}.
}
\description{
Calculates the required sample size for a target power using an
analytic formula based on the methods of Tian et al. (2014).
}
\note{
This function is analytic and does not use bootstrapping.
}
