% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent_censoring_analytical_app.R
\name{.estimate_dependent_censoring_params}
\alias{.estimate_dependent_censoring_params}
\title{Internal Helper to Estimate Dependent Censoring Model Parameters}
\usage{
.estimate_dependent_censoring_params(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  dep_cens_status_var,
  linear_terms,
  L
)
}
\value{
A list containing \code{beta_effect} (the estimated treatment effect) and
\code{se_beta_n1} (the standard error for a sample size of 1).
}
\description{
This internal function contains the common logic for estimating the
treatment effect and its variance from pilot data for the dependent censoring model.
}
\keyword{internal}
