% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicative_stratified.R
\name{.get_internal_simulation_runner}
\alias{.get_internal_simulation_runner}
\title{Internal Factory for RMST Simulation Functions}
\usage{
.get_internal_simulation_runner()
}
\value{
A function that takes \code{n_per_stratum} and runs the simulation.
}
\description{
This is a non-exported helper function that sets up and returns
another function responsible for running the core simulation logic. This
factory pattern helps reduce code duplication between the public-facing
power and sample size functions.
}
\keyword{internal}
