% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rmst-datasets.R
\docType{data}
\name{ph_pwexp_L18_n250}
\alias{ph_pwexp_L18_n250}
\title{Simulated dataset: PH piecewise-exponential, L = 18, n = 250}
\format{
A data frame with 250 rows and 5 variables.
}
\description{
Columns: time, status, arm, age, gender.
}
\examples{
\dontrun{
data(ph_pwexp_L18_n250)
summary(ph_pwexp_L18_n250$time)
}
}
\keyword{datasets}
