% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rmst-datasets.R
\docType{data}
\name{ph_weibull_L24_n300}
\alias{ph_weibull_L24_n300}
\title{Simulated dataset: PH Weibull, L = 24, n = 300}
\format{
A data frame with 300 rows and 5 variables.
}
\description{
Columns: time, status, arm, age, gender.
}
\examples{
\dontrun{
data(ph_weibull_L24_n300)
prop.table(table(ph_weibull_L24_n300$status))
}
}
\keyword{datasets}
