% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_gam.R
\name{design_rmst_gam_power}
\alias{design_rmst_gam_power}
\title{Analyze Power for a Multiplicative Stratified RMST Model (Analytic)}
\usage{
design_rmst_gam_power(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  strata_var,
  sample_sizes,
  linear_terms = NULL,
  tau,
  alpha = 0.05
)
}
\arguments{
\item{pilot_data}{A data.frame with pilot study data.}

\item{time_var, status_var, arm_var, strata_var}{Strings for column names.}

\item{sample_sizes}{A numeric vector of sample sizes to calculate power for.}

\item{linear_terms}{Optional vector of covariates for the model.}

\item{tau}{The truncation time for RMST.}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list containing \code{results_data} and \code{results_plot}.
}
\description{
Performs power analysis using a formula based on the asymptotic
variance from Wang et al. (2019) for a multiplicative, stratified RMST model.
}
\note{
This function is analytic. To simplify the complex estimation procedure
in the paper, it uses a weighted log-linear model on the pilot data to
approximate the effect size and its variance. \code{arm_var} should be a 0/1 indicator.
}
\references{
Wang, X., et al. (2019). Computationally efficient inference for
center effects based on restricted mean survival time. \emph{Statistics in Medicine}, 38, 5133-5145.
}
