% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_gam.R
\name{design_rmst_gam_ss}
\alias{design_rmst_gam_ss}
\title{Find Sample Size for a Semiparametric Additive RMST Model}
\usage{
design_rmst_gam_ss(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  strata_var = NULL,
  target_power,
  linear_terms = NULL,
  smooth_terms = NULL,
  tau,
  n_sim = 1000,
  alpha = 0.05,
  parallel.cores = 1,
  patience = 5,
  n_start = 50,
  n_step = 25,
  max_n_per_arm = 2000
)
}
\arguments{
\item{pilot_data}{A data.frame with pilot study data.}

\item{time_var, status_var, arm_var}{Strings for column names.}

\item{strata_var}{An optional string for a stratification variable.}

\item{target_power}{A single numeric value for the target power.}

\item{linear_terms}{Optional character vector of covariates with a linear effect.}

\item{smooth_terms}{Optional character vector of covariates with a non-linear effect.}

\item{tau}{The truncation time for RMST.}

\item{n_sim}{Number of bootstrap simulations per search step.}

\item{alpha}{The significance level.}

\item{parallel.cores}{Number of cores for parallel processing. Default is 1 (sequential).}

\item{patience}{Number of consecutive non-improving steps in the iterative
search before terminating. Default is 5.}

\item{n_start}{The starting sample size per arm/stratum for the search.}

\item{n_step}{The increment in sample size at each step of the search.}

\item{max_n_per_arm}{The maximum sample size per arm/stratum to search up to.}
}
\value{
A list containing results data, a ggplot object of the search path,
and a summary data.frame.
}
\description{
Performs an iterative sample size search to achieve a target power
using a flexible, semiparametric additive model for the RMST.
}
\note{
\code{status_var} should be \code{1} for an event, \code{0} for censored. \code{arm_var}
should be \code{1} for treatment, \code{0} for control.
}
\examples{
# Create a small pilot dataset for the example
pilot_data_gam <- data.frame(
  surv_time = c(stats::rexp(50, 0.1), stats::rexp(50, 0.05)), #<-- effect
  censor_status = stats::rbinom(100, 1, 0.8),
  arm = rep(0:1, each = 50),
  age = stats::rnorm(100, 50, 10)
)

# --- Sample Size Calculation Example ---
results_n <- design_rmst_gam_ss(
  pilot_data = pilot_data_gam,
  time_var = "surv_time",
  status_var = "censor_status",
  arm_var = "arm",
  target_power = 0.80,
  linear_terms = c("age"),
  tau = 10,
  n_sim = 20, # Low n_sim for example speed
  n_step = 100,
  patience = 2
)
}
\references{
This function uses a pseudo-observation approach. For an alternative
IPCW-based method, see Zhang, Y., & Schaubel, D. E. (2024), Biometrical Journal.
}
