% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ipcw.R
\name{design_rmst_ipcw_power}
\alias{design_rmst_ipcw_power}
\title{Analyze Power for a Linear RMST Model (Analytic)}
\usage{
design_rmst_ipcw_power(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  sample_sizes,
  linear_terms = NULL,
  tau,
  alpha = 0.05
)
}
\arguments{
\item{pilot_data}{A data.frame with pilot study data.}

\item{time_var, status_var, arm_var}{Strings for column names.}

\item{sample_sizes}{A numeric vector of sample sizes per arm to calculate power for.}

\item{linear_terms}{Optional vector of covariates for the linear model.}

\item{tau}{The truncation time for RMST.}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list with \code{results_data} (a data.frame of sample sizes and powers)
and \code{results_plot} (a ggplot object).
}
\description{
Performs power analysis using a direct formula based on the
asymptotic sandwich variance estimator for the linear RMST model (Tian et al., 2014).
}
\note{
\code{status_var} should be \code{1} for an event, \code{0} for censored. \code{arm_var}
should be \code{1} for treatment, \code{0} for control. This function is analytic and
does not use bootstrapping.
}
\references{
Tian, L., et al. (2014). \emph{Biostatistics}, 15(2), 222-233.
}
