% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ipcw.R
\name{design_rmst_ipcw_ss}
\alias{design_rmst_ipcw_ss}
\title{Find Sample Size for a Linear RMST Model}
\usage{
design_rmst_ipcw_ss(
  pilot_data,
  time_var,
  status_var,
  arm_var,
  target_power,
  linear_terms = NULL,
  tau,
  n_sim = 1000,
  alpha = 0.05,
  patience = 5,
  n_start = 50,
  n_step = 25,
  max_n_per_arm = 2000
)
}
\arguments{
\item{pilot_data}{A data.frame with pilot study data.}

\item{time_var, status_var, arm_var}{Strings for column names.}

\item{target_power}{A single numeric value for the target power.}

\item{linear_terms}{Optional vector of covariates for the linear model.}

\item{tau}{The truncation time for RMST.}

\item{n_sim}{Number of bootstrap simulations per search step.}

\item{alpha}{The significance level.}

\item{patience}{Number of consecutive non-improving steps in the iterative
search before terminating. Default is 5.}

\item{n_start}{The starting sample size per arm for the search.}

\item{n_step}{The increment in sample size at each step of the search.}

\item{max_n_per_arm}{The maximum sample size per arm to search up to.}
}
\value{
A list containing \code{results_data} (the final required N), a ggplot object
of the search path, and a \code{results_summary} data.frame.
}
\description{
Performs an iterative sample size search to achieve a target power
based on the direct linear regression model for RMST.
}
\note{
\code{status_var} should be \code{1} for an event, \code{0} for censored. \code{arm_var}
should be \code{1} for treatment, \code{0} for control.
}
\examples{
# Use the 'veteran' dataset for a stable example
data(veteran)
veteran_pilot <- veteran \%>\%
  dplyr::mutate(
    status = ifelse(status == 1, 1, 0),
    trt = ifelse(trt == 2, 1, 0)
  ) \%>\%
  tidyr::drop_na(time, status, trt, karno, age)

results_n <- design_rmst_ipcw_ss(
  pilot_data = veteran_pilot,
  time_var = "time",
  status_var = "status",
  arm_var = "trt",
  target_power = 0.80,
  linear_terms = c("age", "karno"),
  tau = 365,
  n_sim = 50, # Low n_sim for example speed
  patience = 3,
  n_step = 100
)
}
