% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the RMSTdesign Shiny Application}
\usage{
run_app()
}
\description{
A function to easily launch the interactive Shiny application bundled with
the package.
}
\examples{
\dontrun{
  run_app()
}
}
